port harmless fixes over from master, especially 'fix wrong keyboard characters in dialogs for non QWERTY keyboards'

+++ CrashHandler.cpp
@@ -179,7 +179,7 @@ namespace
       {
          fprintf(f, "Attempt to %s 0x%08X\n",
             (exceptionPtrs->ExceptionRecord->ExceptionInformation[0] == 1 ?
-            "write to" : "read from"), exceptionPtrs->ExceptionRecord->ExceptionInformation[1]);
+            "write to" : "read from"), (unsigned int) exceptionPtrs->ExceptionRecord->ExceptionInformation[1]);
       }
       const DWORD threadId = ::GetCurrentThreadId();
       fprintf(f, "Thread ID: 0x%X [%lu]\n\n", threadId, threadId);
@@ -235,14 +235,10 @@ namespace
    void WriteMemoryStatus(FILE* f, const rde::MemoryStatus& status)
    {
       fprintf(f, "Memory status\n=============\n");
-      fprintf(f, "Total Reserved: %uK (%uM) bytes\n", status.totalReserved >> 10,
-         status.totalReserved >> 20);
-      fprintf(f, "Total Commited: %uK (%uM) bytes\n", status.totalCommited >> 10,
-         status.totalCommited >> 20);
-      fprintf(f, "Total Free: %uK (%uM) bytes\n", status.totalFree >> 10,
-         status.totalFree >> 20);
-      fprintf(f, "Largest Free: %uK (%uM) bytes\n\n", status.largestFree >> 10,
-         status.largestFree >> 20);
+      fprintf(f, "Total Reserved: %uK (%uM) bytes\n", (unsigned int)status.totalReserved >> 10, (unsigned int)status.totalReserved >> 20);
+      fprintf(f, "Total Commited: %uK (%uM) bytes\n", (unsigned int)status.totalCommited >> 10, (unsigned int)status.totalCommited >> 20);
+      fprintf(f, "Total Free: %uK (%uM) bytes\n", (unsigned int)status.totalFree >> 10, (unsigned int)status.totalFree >> 20);
+      fprintf(f, "Largest Free: %uK (%uM) bytes\n\n", (unsigned int)status.largestFree >> 10, (unsigned int)status.largestFree >> 20);
    }
 
    void WriteRegisters(FILE* f, const EXCEPTION_POINTERS* exceptionPtrs)
@@ -259,9 +255,9 @@ namespace
       fprintf(f, "RAX=%08X RBX=%08X RCX=%08X RDX=%08X\n" \
          "RSI=%08X RDI=%08X RBP=%08X RSP=%08X RIP=%08X\n" \
          "FLG=%08X CS=%04X DS=%04X SS=%04X ES=%04X FS=%04X GS=%04X\n\n",
-         ctx->Rax, ctx->Rbx, ctx->Rcx, ctx->Rdx, ctx->Rsi, ctx->Rdi,
-         ctx->Rbp, ctx->Rsp, ctx->Rip, ctx->EFlags, ctx->SegCs,
-         ctx->SegDs, ctx->SegSs, ctx->SegEs, ctx->SegFs, ctx->SegGs);
+         (unsigned int)ctx->Rax, (unsigned int)ctx->Rbx, (unsigned int)ctx->Rcx, (unsigned int)ctx->Rdx, 
+         (unsigned int)ctx->Rsi, (unsigned int)ctx->Rdi, (unsigned int)ctx->Rbp, (unsigned int)ctx->Rsp, (unsigned int)ctx->Rip,
+         (unsigned int)ctx->EFlags, (unsigned int)ctx->SegCs, (unsigned int)ctx->SegDs, (unsigned int)ctx->SegSs, (unsigned int)ctx->SegEs, (unsigned int)ctx->SegFs, (unsigned int)ctx->SegGs);
 #else
       fprintf(f, "EAX=%08X EBX=%08X ECX=%08X EDX=%08X\n" \
          "ESI=%08X EDI=%08X EBP=%08X ESP=%08X EIP=%08X\n" \
