+++ vector.h
@@ -12,7 +12,7 @@ public:
    inline void clear() { m_rg.clear(); }
    inline int size() const { return (int)m_rg.size(); }
    inline int capacity() const { return (int)m_rg.capacity(); }
-   inline void resize(const int i) { m_rg.resize(i); }
+   inline void resize(const size_t i) { m_rg.resize(i); }
    inline void push_back(void* const pItem) { m_rg.push_back(pItem); }
 
    inline int find(const void * const pvItem) const
@@ -195,7 +195,7 @@ public:
     {
         if (m_size == m_elements.size())
         {
-            m_indices.insert(m_indices.begin() + n, m_size);
+            m_indices.insert(m_indices.begin() + n, (unsigned int)m_size);
             m_elements.push_back(value);
         }
         else
@@ -218,7 +218,7 @@ public:
     {
         if (m_size == m_elements.size())
         {
-            m_indices.push_back(m_size);
+            m_indices.push_back((unsigned int)m_size);
             m_elements.push_back(value);
         }
         else
